endpoints: support packages and channels
Required in order to support attenuated macaroons that work with a set
of specific packages or channels.

A higher level Package class is introduced to simplify the
presentation of StoreClient's login method.

A RuntimeError is raised to ease development when improper package
types are used for each endpoint (Snap Store and Charmhub).

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>