store client: split common StoreClient functionality to a base class
The BaseClient is reusable between the Candid flow and the Ubuntu One
flow. Required implementation is done with an AbstractBaseClass and
abstractmethods to ensure the implementor does the right thing.

The BaseClient now also composes the HTTPClient instead of inheriting
from it to make it easier for subclassess to make unauthenticated http
calls without needing to change the signature for .request.

The overrides package was addedd to help pydocstyle not request
docstrings for overriden classes which also triggered pylint's
duplicate-code check (which is in any case disabled as the bigger
signatures in the methods also trigger it).

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>