auth: update stored authentication format
Instead of storing a payload that depends on the auth 'flavor' (Candid
or UbuntuOne), store, as a JSON string, a dict containing the token
"type" t and the actual payload value v. Add some
backwards-compatibility support to allow reading credentials stored in
the previous (type-dependent) way. Tagging the token type allows us to
extend supported types in the future.