feat: Don't use error list as details (#295)
A recent-ish change in craft-cli made an error's 'details' get shown to the
user by default. This means that CraftStoreErrors with a single store error
would get the actual error message printed twice - once from the 'brief' and
once from the 'details'. Instead, log the error list (as it has relevant
error codes).